+++ sdk/appplatform/azure-mgmt-appplatform/generated_samples/build_service_agent_pool_update_put.py
@@ -39,6 +39,6 @@ def main():
     print(response)
 
 
-# x-ms-original-file: specification/appplatform/resource-manager/Microsoft.AppPlatform/stable/2022-12-01/examples/BuildServiceAgentPool_UpdatePut.json
+# x-ms-original-file: specification/appplatform/resource-manager/Microsoft.AppPlatform/stable/2023-12-01/examples/BuildServiceAgentPool_UpdatePut.json
 if __name__ == "__main__":
     main()