+++ sdk/keyvault/azure-keyvault-secrets/azure/keyvault/secrets/aio/_client.py
@@ -300,12 +300,17 @@ class SecretClient(AsyncKeyVaultClientBase):
         polling_interval = kwargs.pop("_polling_interval", None)
         if polling_interval is None:
             polling_interval = 2
-        deleted_secret = DeletedSecret._from_deleted_secret_bundle(
-            await self._client.delete_secret(self.vault_url, name, **kwargs)
+        pipeline_response, deleted_secret_bundle = await self._client.delete_secret(
+            vault_base_url=self.vault_url,
+            secret_name=name,
+            cls=lambda pipeline_response, deserialized, _: (pipeline_response, deserialized),
+            **kwargs,
         )
+        deleted_secret = DeletedSecret._from_deleted_secret_bundle(deleted_secret_bundle)
 
         polling_method = AsyncDeleteRecoverPollingMethod(
             # no recovery ID means soft-delete is disabled, in which case we initialize the poller as finished
+            pipeline_response=pipeline_response,
             command=partial(self.get_deleted_secret, name=name, **kwargs),
             final_resource=deleted_secret,
             finished=deleted_secret.recovery_id is None,
@@ -415,13 +420,21 @@ class SecretClient(AsyncKeyVaultClientBase):
         polling_interval = kwargs.pop("_polling_interval", None)
         if polling_interval is None:
             polling_interval = 2
-        recovered_secret = SecretProperties._from_secret_bundle(
-            await self._client.recover_deleted_secret(self.vault_url, name, **kwargs)
+        pipeline_response, recovered_secret_bundle = await self._client.recover_deleted_secret(
+            vault_base_url=self.vault_url,
+            secret_name=name,
+            cls=lambda pipeline_response, deserialized, _: (pipeline_response, deserialized),
+            **kwargs,
         )
+        recovered_secret = SecretProperties._from_secret_bundle(recovered_secret_bundle)
 
         command = partial(self.get_secret, name=name, **kwargs)
         polling_method = AsyncDeleteRecoverPollingMethod(
-            command=command, final_resource=recovered_secret, finished=False, interval=polling_interval
+            pipeline_response=pipeline_response,
+            command=command,
+            final_resource=recovered_secret,
+            finished=False,
+            interval=polling_interval
         )
         await polling_method.run()