+++ sdk/core/azure-core/CHANGELOG.md
@@ -1,14 +1,23 @@
 # Release History
 
-## 1.29.7 (Unreleased)
+## 1.29.8 (Unreleased)
 
 ### Features Added
+
+- Support tuple input for file values  to `azure.core.rest.HttpRequest`  #33948
+- Support tuple input to `files` with duplicate field names  `azure.core.rest.HttpRequest`  #34021
 
 ### Breaking Changes
 
 ### Bugs Fixed
 
 ### Other Changes
+
+## 1.29.7 (2024-01-18)
+
+### Other Changes
+
+- Removed dependency on `anyio`.  #33282
 
 ## 1.29.6 (2023-12-14)